+++ lib/common/word.dart
@@ -1,3 +1,5 @@
+import 'reference.dart';
+
 class HebrewGreekWord {
   /// ID is in the form of BBCCCVVVWW,
   /// where BB is the book number,
@@ -14,12 +16,10 @@ class HebrewGreekWord {
   String toString() => 'HebrewGreekWord(id: $id, text: $text)';
 }
 
-(int bookId, int chapter, int verse, int word) extractReferenceFromWordId(
-  int wordId,
-) {
-  final word = wordId % 100;
+Reference extractReferenceFromWordId(int wordId) {
+  // final word = wordId % 100;
   final verse = (wordId ~/ 100) % 1000;
   final chapter = (wordId ~/ 100000) % 1000;
   final bookId = wordId ~/ 100000000;
-  return (bookId, chapter, verse, word);
+  return Reference(bookId: bookId, chapter: chapter, verse: verse);
 }