+++ lib/ui/search/search.dart
@@ -1,4 +1,6 @@
 import 'package:flutter/material.dart';
+import 'package:studyapp/common/book_name.dart';
+import 'package:studyapp/common/reference.dart';
 import 'package:studyapp/l10n/app_localizations.dart';
 import 'package:studyapp/ui/search/keyboard.dart';
 
@@ -25,7 +27,7 @@ class _SearchPageState extends State<SearchPage> {
 
   void _onTextChanged() {
     final prefix = _controller.text;
-    manager.search(prefix);
+    manager.searchWordPrefix(prefix);
   }
 
   @override
@@ -58,7 +60,7 @@ class _SearchPageState extends State<SearchPage> {
               ),
             ),
             Expanded(
-              child: ValueListenableBuilder(
+              child: ValueListenableBuilder<SearchResults>(
                 valueListenable: manager.resultsNotifier,
                 builder: (context, results, child) {
                   return Directionality(
@@ -66,13 +68,72 @@ class _SearchPageState extends State<SearchPage> {
                     child: ListView.builder(
                       itemCount: results.length,
                       itemBuilder: (context, index) {
-                        final word = results[index];
-                        return ListTile(
-                          title: Text(
-                            word,
-                            style: TextStyle(fontFamily: 'sbl'),
-                          ),
-                        );
+                        switch (results) {
+                          case NoResults():
+                            return const SizedBox();
+                          case WordSearchResults(words: final w):
+                            final word = w[index];
+                            return ListTile(
+                              title: Text(
+                                word,
+                                style: TextStyle(fontFamily: 'sbl'),
+                              ),
+                              onTap: () {
+                                manager.searchVerses(word);
+                              },
+                            );
+                          case VerseSearchResults(
+                            searchWord: final word,
+                            references: final r,
+                          ):
+                            final fontSize = 30.0;
+                            final reference = r[index];
+                            final formattedReference = _formatReference(
+                              reference,
+                            );
+                            return FutureBuilder<TextSpan>(
+                              future: manager.getVerseContent(
+                                word,
+                                reference,
+                                Theme.of(context).colorScheme.primary,
+                                fontSize,
+                              ),
+                              builder: (context, snapshot) {
+                                if (snapshot.hasError) {
+                                  return ListTile(
+                                    title: Text(formattedReference),
+                                    subtitle: Text(
+                                      'Error loading verse: ${snapshot.error}',
+                                    ),
+                                  );
+                                }
+                                if (snapshot.connectionState ==
+                                    ConnectionState.done) {
+                                  final verse = snapshot.data!;
+                                  return ListTile(
+                                    title: Text(
+                                      formattedReference,
+                                      style: TextStyle(
+                                        fontFamily: 'sbl',
+                                        fontSize: fontSize,
+                                        color: Theme.of(context).disabledColor,
+                                      ),
+                                    ),
+                                    subtitle: Text.rich(
+                                      verse,
+                                      textDirection: _textDirection,
+                                    ),
+                                  );
+                                } else {
+                                  // Giving the widget a height ensures that the
+                                  // ListView.builder will not try to build the
+                                  // every item in the list just because they all
+                                  // theoretically fit with a zero height.
+                                  return const SizedBox(height: 50);
+                                }
+                              },
+                            );
+                        }
                       },
                     ),
                   );
@@ -94,5 +155,10 @@ class _SearchPageState extends State<SearchPage> {
         ),
       ),
     );
+  }
+
+  String _formatReference(Reference reference) {
+    final bookName = bookNameForId(context, reference.bookId);
+    return '$bookName ${reference.chapter}:${reference.verse}';
   }
 }